backend/requirements.txt
uvicorn
requests
fastapi
pydantic
selenium==4.10.0
pytz
pyyaml
Selenium-Screenshot>=2.1.0
pytest
loguru